fix server
Merge debug

+++ Projekt/include/statistics.h
@@ -3,7 +3,7 @@
 
 #include "common.h"
 
-float calculate_packet_loss(int packets_received, int packets_sent);
-void print_internet_speed(float packet_loss, int how_many_bytes);
+double calculate_packet_loss(double packets_received, double packets_sent);
+void print_internet_speed(double packet_loss, int how_many_bytes);
 
 #endif
